+++ core/blockchain.go
@@ -1775,7 +1775,6 @@ func (bc *BlockChain) insertChain(chain types.Blocks, setHead bool) (int, error)
 
 		// Checking if the headers contains at least one MontBlanc block.
 		if bc.chainConfig.IsMontBlanc(headers[len(headers)-1].Number) || bc.chainConfig.MontBlancBlock == nil {
-
 			for _, header := range headers {
 				// Checking if it is a MontBlanc Block or a new chain.
 				if bc.chainConfig.IsMontBlanc(header.Number) || bc.chainConfig.MontBlancBlock == nil {
